rustpkg Initialize build-context later in the build process
